+++ includes/features/class-schema.php
@@ -371,11 +371,9 @@ class Schema {
 	public static function get_grouped_list( $filter, $group = '', $cache = true, $extra_field = '', $extras = [], $not = false, $order = '', $limit = 0 ) {
 		// phpcs:ignore
 		$id = Cache::id( __FUNCTION__ . serialize( $filter ) . $group . $extra_field . serialize( $extras ) . ( $not ? 'no' : 'yes') . $order . (string) $limit);
-		if ( $cache ) {
-			$result = Cache::get_global( $id );
-			if ( $result ) {
-				return $result;
-			}
+		$result = Cache::get_global( $id );
+		if ( $result ) {
+			return $result;
 		}
 		if ( '' !== $group ) {
 			$group = ' GROUP BY ' . $group;
@@ -389,9 +387,7 @@ class Schema {
 		// phpcs:ignore
 		$result = $wpdb->get_results( $sql, ARRAY_A );
 		if ( is_array( $result ) ) {
-			if ( $cache ) {
-				Cache::set_global( $id, $result, 'infinite' );
-			}
+			Cache::set_global( $id, $result, $cache ? 'infinite' : 'ephemeral' );
 			return $result;
 		}
 		return [];