+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to **Sessions** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Sessions** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.0] - Not Yet Released
+
+### Changed
+- Improved ephemeral cache in analytics.
+
 ## [2.7.0] - 2022-04-20
 
 ### Added